added interval controller
