+++ FoqosDeviceMonitor/Info.plist
