--- foqosblockedtracker/DeviceActivityMonitorExtension.swift
@@ -1,53 +1,0 @@
-import DeviceActivity
-import ManagedSettings
-
-// Optionally override any of the functions below.
-// Make sure that your class name matches the NSExtensionPrincipalClass in your Info.plist.
-class DeviceActivityMonitorExtension: DeviceActivityMonitor {
-    let store = ManagedSettingsStore(named: ManagedSettingsStore.Name("foqosAppRestrictions"))
-    
-    override func intervalDidStart(for activity: DeviceActivityName) {
-        super.intervalDidStart(for: activity)
-        
-        if activity == DeviceActivityName("foqosDeviceActivity") {
-            if let blockedApps = store.application.blockedApplications {
-                let blockedTokens = blockedApps.compactMap { $0.token }
-                store.shield.applications = Set(blockedTokens)
-            } else {
-                store.shield.applications = nil
-            }
-        }
-    }
-    
-    override func intervalDidEnd(for activity: DeviceActivityName) {
-        super.intervalDidEnd(for: activity)
-        
-        if activity == DeviceActivityName("FoqosDeviceActivity") {
-            store.shield.applications = nil
-        }
-    }
-    
-    override func eventDidReachThreshold(_ event: DeviceActivityEvent.Name, activity: DeviceActivityName) {
-        super.eventDidReachThreshold(event, activity: activity)
-        
-        // Handle the event reaching its threshold.
-    }
-    
-    override func intervalWillStartWarning(for activity: DeviceActivityName) {
-        super.intervalWillStartWarning(for: activity)
-        
-        // Handle the warning before the interval starts.
-    }
-    
-    override func intervalWillEndWarning(for activity: DeviceActivityName) {
-        super.intervalWillEndWarning(for: activity)
-        
-        // Handle the warning before the interval ends.
-    }
-    
-    override func eventWillReachThresholdWarning(_ event: DeviceActivityEvent.Name, activity: DeviceActivityName) {
-        super.eventWillReachThresholdWarning(event, activity: activity)
-        
-        // Handle the warning before the event reaches its threshold.
-    }
-}
